refactor(store): extract _SchemaManagerMixin to _schema.py (#280)
Second step of the VstashStore decomposition (#280). Move-only: the 8
schema-concern methods (_create_tables, _check_schema_version,
_migrate_v1_to_v2, _ddl_declares_cosine, _migrate_schema, get_meta, set_meta,
check_embedding_drift) move verbatim into _SchemaManagerMixin in
vstash/store/_schema.py; VstashStore inherits it. _connect stays on the facade
(it builds the connection and __init__ calls it).

The only non-blank changes to __init__.py are the mixin import and the class
base. Method bodies are byte-identical to the originals (verified via diff).

- pytest tests/ -> 1248 passed (1244 + 4 new test_schema_manager.py), 6 deselected
- ruff check/format clean
